Move tests for validateDobForGro from BirthEventsControllerDobSwitchSpec into BirthEventsControllerSpec
